Update README.md
First git public release of AUTOPAR2.CMD, first written in 2004, with limited availability to testing on usenet groups. Limited customization, more of overall functionality to check PARs and RARs from various usenet groups . This was prior to the dldone, and rarslave, Par Buddy.  Most of the functionality in AUTOPAR2.CMD is built into most modern day download programs, such as Sabnzbd and nzbGET. This is one of the pioneer scripts to make usenet downloads less intensive on the user, and to make the checking of Newzleecher,  GrabIT, and Newsbin Pro articles more streamlined.